% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdep-package.R
\docType{package}
\name{ncdep-package}
\alias{ncdep}
\alias{ncdep-package}
\alias{ncdep-package}
\alias{ncdep}
\title{ncdep: Test NetCDF on Travis and Appveyor}
\description{
What the package does (one paragraph).

NetCDF dependencies
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
