% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdep-package.R
\docType{data}
\name{oceandata_dap}
\alias{oceandata_dap}
\title{NASA ocean colour example}
\description{
The URL of an ocean colour NetCDF file behind a Thredds server.
}
\examples{
x1 <- ncdf4::nc_open(oceandata_dap)
x2 <- RNetCDF::open.nc(oceandata_dap)
}
