% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdep-package.R
\docType{data}
\name{unidata_file}
\alias{unidata_file}
\title{Unidata example file}
\description{
The relative path of the Unidata madis-hydro example file.
}
\examples{
madishydro <- system.file(unidata_file, package = "ncdep")
file.exists(madishydro)
}
